+++ ai-logic/firebase-ai/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- [feature] Added static factory methods `createWithThinking` for `Part` implementations to expose `thoughtSignature` and `isThought` properties. (#8352)
 - [feature] Added automatic function calling support for `LiveGenerativeModel`. (#8223)
 
 # 17.13.0